pull flawed overlay work
unfortunately the gpio-poweroff[0] overlay sets the pin HI->LO->HI,
which is not compatible with the x-c1 expansion board; it expects the
pin to end on LO. I can't figure out how to do that without patching the
kernel. a long/infinite active_delay_ms or inactive_delay_ms keeps some
part of the kernel alive holding the pin and it never releases.

the behavior observed is that the system will power down then
briefly power back up before powering down for good. and you have to
press the power button twice to get it to boot again. no bueno...

any shutdowns initiated from the os -- including by the gpio-shutdown
overlay -- need to be able to correctly cut power at the right point
before the CPU/SoC reset. so gpio-poweroff effectively becomes a
prequisite of gpio-shutdown on this platform. this is probably why
geekworm had to come up with the janky xPWR script...

[0]: https://www.kernel.org/doc/Documentation/devicetree/bindings/power/reset/gpio-poweroff.txt